+++ RxCocoa/Traits/SharedSequence/SharedSequence+Operators.swift
@@ -493,3 +493,20 @@ extension SharedSequenceConvertibleType {
         return SharedSequence(source)
     }
 }
+
+// MARK: elementAt
+extension SharedSequenceConvertibleType {
+
+    /**
+     Returns a sequence emitting only element _n_ emitted by an Observable
+
+     - seealso: [elementAt operator on reactivex.io](http://reactivex.io/documentation/operators/elementat.html)
+
+     - parameter index: The index of the required element (starting from 0).
+     - returns: An observable sequence that emits the desired element as its own sole emission.
+     */
+    public func elementAt(_ index: Int) -> SharedSequence<SharingStrategy, E> {
+        let source = asObservable().elementAt(index)
+        return SharedSequence(source)
+    }
+}