+++ Tests/RxCocoaTests/Driver+Test.swift
@@ -1388,3 +1388,99 @@ extension DriverTest {
         }
     }
 }
+
+// MARK: elementAt
+
+extension DriverTest {
+    func testElementAt_CompleteAfter() {
+        let scheduler = TestScheduler(initialClock: 0)
+
+        driveOnScheduler(scheduler) {
+            let observables = scheduler.createHotObservable([
+                next(70, 6),
+                next(150, 4),
+                next(210, 9),
+                next(230, 13),
+                next(270, 7),
+                next(280, 1),
+                next(300, -1),
+                next(310, 3),
+                next(340, 8),
+                next(370, 11),
+                next(410, 15),
+                next(415, 16),
+                next(460, 72),
+                next(510, 76),
+                next(560, 32),
+                next(570, -100),
+                next(580, -3),
+                next(590, 5),
+                next(630, 10),
+                completed(690)
+            ])
+
+            let res = scheduler.start {
+                observables.elementAt(10).asObservable()
+            }
+
+            XCTAssertEqual(res.events, [
+                next(460, 72),
+                completed(460)
+            ])
+        }
+    }
+
+    func testElementAt_OutOfRangeError() {
+        let scheduler = TestScheduler(initialClock: 0)
+
+        driveOnScheduler(scheduler) {
+            let observables = scheduler.createHotObservable([
+                next(70, 6),
+                next(150, 4),
+                next(210, 9),
+                next(230, 13),
+                next(270, 7),
+                next(280, 1),
+                next(300, -1),
+                completed(320)
+            ])
+
+            let res = scheduler.start {
+                observables.elementAt(10).asObservable()
+            }
+
+            XCTAssertEqual(res.events, [
+                error(320, RxError.argumentOutOfRange)
+            ])
+        }
+    }
+
+    func testElementAt_FirstElement() {
+        let scheduler = TestScheduler(initialClock: 0)
+
+        driveOnScheduler(scheduler) {
+            let observables = scheduler.createHotObservable([
+                next(70, 6),
+                next(150, 4),
+                next(210, 9),
+                next(230, 13),
+                next(270, 7),
+                next(280, 1),
+                next(300, -1),
+                next(310, 3),
+                next(340, 8),
+                next(370, 11),
+                completed(400)
+            ])
+
+            let res = scheduler.start {
+                observables.elementAt(0).asObservable()
+            }
+
+            XCTAssertEqual(res.events, [
+                next(210, 9),
+                completed(210)
+            ])
+        }
+    }
+}