+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## Master
 
-* Adds `from` operator on "SharedSequence"
+* Adds `from` and `elementAt` operator on "SharedSequence"
 * Adds `merge` operator on "Completable"
 * Adds `using` operator on "PrimitiveSequence"